Compile status atom refactor changes
