+++ plugin/src/features/Deployment/index.tsx
@@ -7,7 +7,7 @@ import { useAtom, useAtomValue, useSetAtom } from 'jotai'
 import CompiledContracts from '@/components/CompiledContracts'
 import './styles.css'
 import Container from '../../ui_components/Container'
-import { type AccordianTabs } from '../Plugin'
+import { type AccordianTabs } from '@/types/common'
 import ConstructorInput from '../../components/ConstructorInput'
 import { type VerificationResult, type DeployedContract, type ContractFile } from '../../types/contracts'
 import { mockManualChain, type Transaction } from '../../types/transaction'
