+++ plugin/src/types/common.ts
@@ -1,0 +1,1 @@
+export type AccordianTabs = 'compile' | 'deploy' | 'interaction' | 'transactions' | ''
